query-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
firebase-admin==6.8.0
pydantic==2.5.2
openai>=1.10.0,<2.0.0
python-dotenv==1.0.0 